Fix bbox to angle estimation
Yolo predictions are xyxy-absolute.

+++ src/smirk/pedestrian_detector/pedestrian_detector.py
@@ -7,10 +7,10 @@ import numpy as np
 
 @dataclass
 class BoundingBox:
-    x_min: float
-    y_min: float
-    x_max: float
-    y_max: float
+    x_min: int
+    y_min: int
+    x_max: int
+    y_max: int
 
 
 class PedestrianDetector(ABC):
